+++ src/structure/mod.rs
@@ -1,5 +1,4 @@
 pub mod editor;
-pub(crate) mod protocol;
 pub mod utils;
 pub(crate) mod packet;
