+++ src/structure/utils/traits/mod.rs
@@ -3,3 +3,4 @@ pub(crate)mod TypeManager;
 pub(crate)mod Detector;
 pub(crate)mod TrtFrom;
 pub(crate)mod From;
+pub(crate)mod ProtocolCodec;
